requirement.txt
# To activate this run
#
# pip install -r requirements.txt
# python -m django --version
#
# On a Macintosh this should be python3 and pip3
#
django-crispy-forms==1.8.1
django-filter==2.2.0
djangorestframework==3.11.2
debug_toolbar=3.
drf_yasg=1.20.1
Markdown==3.1.1
social-auth-app-django==3.1.0
social-auth-core==3.3.3
django-extensions==2.2.5
django-taggit==1.3.0
six==1.16.0

# If you want to use MySQL, and are not running on PythonAnywhere,
# uncomment the following line you might need to install pre-requisite
# libraries before mysqlclient will work - instructions are at
#   https://pypi.org/project/mysqlclient/
